Game fixed

+++ Dealer/src/main/kotlin/com/abaddon83/burraco/dealer/events/DealerEvent.kt
@@ -2,9 +2,9 @@ package com.abaddon83.burraco.dealer.events
 
 import io.github.abaddon.kcqrs.core.domain.messages.events.IDomainEvent
 
-abstract class DealerEvent: IDomainEvent {
+abstract class DealerEvent : IDomainEvent {
     override val aggregateType: String = AGGREGATE_TYPE_NAME
-    override val version: Int = 1
+    override val version: Long = 1
 
     companion object {
         @JvmStatic
